+++ kepler-core/src/relationships/event_order.rs
@@ -1,6 +1,7 @@
 use super::super::models::*;
 use crate::hash::Hash;
 use crate::relationships::*;
+use crate::types::orbit_id_wrap::OrbitIdWrap;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel, PartialOrd, Ord)]
@@ -14,7 +15,7 @@ pub struct Model {
     pub epoch_seq: i64,
     pub event: Hash,
 
-    pub orbit: epoch::OrbitIdWrap,
+    pub orbit: OrbitIdWrap,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
